requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic[email]>=2.7.1
pydantic-settings>=2.0.0
pytest>=8.2.0
httpx>=0.27.0
